+++ Changelog.md
@@ -1,12 +1,13 @@
-## Changelog
+## Bemto-components changelog
 
 ### v1.6.0 (in development)
 
 - Added a way to pass function to props in definition.
 - Added a way to pass function to `optional` prop at elem definition.
 - Added a way to disable an element by calling it as `false` or `null`.
-- Added docs for the external `bemto-button` component.
+- Added a way to omit tag on element.
 - Added the `reset` as a possible button type.
+- Added docs for the external `bemto-button` component.
 - Fixed non-cloned parents array which led to incorrect render of elements in some cases.
 
 ### v1.5.0 (2017-11-18)